@@ -1,7 +1,7 @@
 // Copyright © 2023 Denis Morel
 
 // This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU Lesser General Public License as published by the Free
+// the terms of the GNU General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option) any
 // later version.
 //
@@ -10,18 +10,18 @@
 // FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
 // details.
 //
-// You should have received a copy of the GNU Lesser General Public License and
+// You should have received a copy of the GNU General Public License and
 // a copy of the GNU General Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/>.
 
 //! Algorithms defined in section Preliminaries
 
-mod election_event_context;
+mod agreement;
 mod electoral_model;
 mod proof_of_correct_key_generation;
 mod write_ins;
 
-pub use election_event_context::*;
+pub use agreement::*;
 pub use electoral_model::*;
 pub use proof_of_correct_key_generation::*;
 pub use write_ins::*;
