@@ -1,7 +1,7 @@
 // Copyright © 2023 Denis Morel
 
 // This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU Lesser General Public License as published by the Free
+// the terms of the GNU General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option) any
 // later version.
 //
@@ -10,11 +10,11 @@
 // FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
 // details.
 //
-// You should have received a copy of the GNU Lesser General Public License and
+// You should have received a copy of the GNU General Public License and
 // a copy of the GNU General Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/>.
 
-use super::WriteInsError;
+use super::{WriteInsError, WriteInsErrorRepr};
 use rust_ev_crypto_primitives::{
     alphabets::ALPHABET_LATIN, elgamal::EncryptionParameters, ConstantsTrait, Integer,
     OperationsTrait,
@@ -31,30 +31,23 @@ pub fn quadratic_residue_to_write_in(
     let q = encryption_parameters.q();
     let mut x = y
         .mod_exponentiate(&(Integer::from(p + 1) / 4), p)
-        .map_err(|e| {
-            WriteInsError::IntegerToWriteInput(format!("Error with mod_exponentiate {}", e))
-        })?;
+        .map_err(|e| WriteInsErrorRepr::QuadraticToWriteIns { source: e })?;
     if &x > q {
         x = p - x;
     }
-    integer_to_write_in(encryption_parameters, &x)
+    integer_to_write_in(encryption_parameters.q(), &x)
 }
 
 /// Algorithm 3.17
 ///
 /// Error [WriteInsError] if something is going wrong
-fn integer_to_write_in(
-    encryption_parameters: &EncryptionParameters,
-    x: &Integer,
-) -> Result<String, WriteInsError> {
+fn integer_to_write_in(q: &Integer, x: &Integer) -> Result<String, WriteInsError> {
     if x <= Integer::zero() {
-        return Err(WriteInsError::IntegerToWriteInput(
-            "x cannot be less or equal 0".to_string(),
-        ));
+        return Err(WriteInsError::from(WriteInsErrorRepr::XPositive));
     }
     let a = ALPHABET_LATIN.size();
     // ensure that x is in Z_q
-    let mut x_internal = Integer::from(x % encryption_parameters.q());
+    let mut x_internal = Integer::from(x % q);
     let mut res = String::new();
     loop {
         if &x_internal == Integer::zero() {
@@ -70,48 +63,28 @@ fn integer_to_write_in(
 
 #[cfg(test)]
 mod test {
-    use std::fs;
-
-    use crate::{
-        test_data::get_test_data_writeins_path,
-        test_json_data::{json_to_encryption_parameters_base16, json_value_to_integer_base16},
-    };
-
     use super::*;
-    use serde_json::Value;
-
-    fn get_test_case_integer_to_write_in() -> Value {
-        serde_json::from_str(
-            &fs::read_to_string(get_test_data_writeins_path().join("integer-to-write-in.json"))
-                .unwrap(),
-        )
-        .unwrap()
-    }
-
-    fn get_test_case_qr_to_write_in() -> Value {
-        serde_json::from_str(
-            &fs::read_to_string(get_test_data_writeins_path().join("qr-to-write-in.json")).unwrap(),
-        )
-        .unwrap()
-    }
+    use crate::{
+        test_data::get_test_data_writeins,
+        test_json_data::{json_to_encryption_parameters_base64, json_value_to_integer_base64},
+    };
 
     #[test]
     fn test_integer_to_write_in() {
-        let value = get_test_case_integer_to_write_in();
+        let value = get_test_data_writeins("integer-to-write-in.json");
         for tc in value.as_array().unwrap().iter() {
             let description = tc["description"].as_str().unwrap();
-            let ep = json_to_encryption_parameters_base16(&tc["context"]);
-            let x = json_value_to_integer_base16(&tc["input"]["x"]);
+            let q = json_value_to_integer_base64(&tc["context"]["q"]);
+            let x = json_value_to_integer_base64(&tc["input"]["x"]);
             let expected = tc["output"]["output"].as_str().unwrap();
             assert_eq!(
-                integer_to_write_in(&ep, &x).unwrap().as_str(),
+                integer_to_write_in(&q, &x).unwrap().as_str(),
                 expected,
-                "{}",
-                description
+                "{description}"
             )
         }
         assert!(integer_to_write_in(
-            &json_to_encryption_parameters_base16(&value[0]["context"]),
+            &json_value_to_integer_base64(&value[0]["context"]["q"]),
             Integer::zero()
         )
         .is_err())
@@ -119,17 +92,16 @@ mod test {
 
     #[test]
     fn test_qr_to_write_in() {
-        let value = get_test_case_qr_to_write_in();
+        let value = get_test_data_writeins("quadratic-residue-to-write-in.json");
         for tc in value.as_array().unwrap().iter() {
             let description = tc["description"].as_str().unwrap();
-            let ep = json_to_encryption_parameters_base16(&tc["context"]);
-            let y = json_value_to_integer_base16(&tc["input"]["y"]);
+            let ep = json_to_encryption_parameters_base64(&tc["context"]);
+            let y = json_value_to_integer_base64(&tc["input"]["y"]);
             let expected = tc["output"]["output"].as_str().unwrap();
             assert_eq!(
                 quadratic_residue_to_write_in(&ep, &y).unwrap().as_str(),
                 expected,
-                "{}",
-                description
+                "{description}"
             )
         }
     }
