@@ -1,7 +1,7 @@
 // Copyright © 2023 Denis Morel
 
 // This program is free software: you can redistribute it and/or modify it under
-// the terms of the GNU Lesser General Public License as published by the Free
+// the terms of the GNU General Public License as published by the Free
 // Software Foundation, either version 3 of the License, or (at your option) any
 // later version.
 //
@@ -10,7 +10,7 @@
 // FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
 // details.
 //
-// You should have received a copy of the GNU Lesser General Public License and
+// You should have received a copy of the GNU General Public License and
 // a copy of the GNU General Public License along with this program. If not, see
 // <https://www.gnu.org/licenses/>.
 
@@ -20,15 +20,20 @@ mod process_plaintexts;
 mod verifiy_client_proofs;
 mod verify_mix_dec_offline;
 
+use crate::preliminaries::ElectoralModelError;
 pub use process_plaintexts::ProcessPlaintextsOutput;
+use thiserror::Error;
 pub use verifiy_client_proofs::*;
 pub use verify_mix_dec_offline::*;
 
-use thiserror::Error;
+/// Errors during the algorithms for Mix Offline
+#[derive(Error, Debug)]
+#[error(transparent)]
+pub struct MixOfflineError(#[from] MixOfflineErrorRepr);
 
 // enum representing the errors during the algorithms for Mix Offline
-#[derive(Error, Debug, Clone)]
-pub enum MixOfflineError {
+#[derive(Error, Debug)]
+pub enum MixOfflineErrorRepr {
     #[error("Error input in ProcessPlaintexts: {0}")]
     ProcessPlaintextsInput(String),
     #[error("Error processing in ProcessPlaintexts: {0}")]
@@ -41,4 +46,6 @@ pub enum MixOfflineError {
     VerifyMixDecOfflineInput(String),
     #[error("Error in processing VerifyMixDecOffline: {0}")]
     VerifyMixDecOfflineProcess(String),
+    #[error("Error getting psi from pTable")]
+    GetPsi { source: ElectoralModelError },
 }
